Create skeleton top-level structure
Previous commits identified what we thought we would need to plot well
trajectories.

This commit implements the top-level code to plot well trajectories. It
has no unit tests, but it runs without throwing an exception. In
addition, this commit implements the main function that expects certain
arguments to be present on the command line.