Add project loader
Previous commits integrated TrajectoryCoordinator with ProjectAdapter
but did not wire up a "project loader" responsible for loading an
`.ifrac` file from disk (into an `IProject`).

This commit "wires" up a ProjectLoader. This action changed the
constructor, TrajectoryCoordinator() which provided an opportunity to
change the TrajectoryCoordinator implementation to use a property for
the loaded project.

In C#, many of these actions would be impacted by a choice of dependency
injection and a container. To preserve architectural flexibility, this
commit *does not* include a DI "container" (either "do-it-yourself" or
from a Python package); however, the author did add a `TODO` comment to
remind our future selves to consider this choice again.