Change project_loader to *not* hard-code timezone
Previous commits hard-coded a `TimeZoneInfo` instance in a call to
`ProjectFileReaderWriterV2.Read`

On later, more detailed, reading, we discovered that the
actual `TimeZoneInfo` argument is *not* used. Further, we observed that
this `Read` overload calls another `Read` overload that *does not*
require a `TimeZoneInfo` insntance.

Note that this change requires copying the code used to initialize the
`Stream` from `ProjectFileReaderWriterV2' to the Python script. Copying
this details fails the "DRY/OAOO" principle ("Don't Repeat Yourself" /
"Once and Only Once"); however, I've added a `TODO` to the code
reminding our future selves of this copy. It is not the best solution,
but I think it is the best way to improve the code in a small way today.

Finally, a third reading of the code discovered that the optional
`bool`, `useExternalProcess`, is never used. Consequently I have removed
setting this parameter in my call.

This commit changes the call from `Read(string, TimeZoneInfo,
bool=false)` to `Read(Stream, bool=false)`.