Write ProjectAdapter.trajectory_points
Previous commits only contained a stub method.

This commit provides a concrete, somewhat tested implementation.

Because of time constraints, I chose to use doctest to test the *actual*
results instead of using unit tests to test predicted results. This kind
of testing mitigates some risk, but not as much risk as unit tests.