Replace StubWells class with mocks
Previous commits introduced a `StubWells` class to be returned from
`loaded_project().Wells`. This worked but introduced an additional
"mocking" mechanism.

This commit removes that class replacing `loaded_project.return_value`
with a `MagicMock` configured to have a chain `Wells.Items` returning
the values appropriate for the test.

CAUTION: a "side-effect" of this change is to make the test code
dependent on the `clr` and the `IProject` type (made available by
`pythonnet` to Python) because I specced the `MagicMock` using
`IProject`. A further consequence of this side-effect seems to be a
slowing down of the tests. For now it is acceptable; however, if too
many tests require this same loading of the dynamic .NET environment,
our future selves may want to remove this dependency.