Write WellsFacade class
Previous commits used the `ProjectAdapter` class as the single place to
adapt the entire object graph reachable from `IProject`.

When prototyping the code to plot (pressure) time series, not only was
the plotting code very similar to the plotting code for trajectories,
but the structure for interacting with `IProject` was very similar to
the `IWell` interaction. Thinking about it, the best course going
forward was to revive our original "coordinator" structure. However, the
"coordinator design pattern" was not quite right, so I chose the GOF
Facade pattern.

This commit prototypes that facade.