Plot all pressure curves (hard-coded)
Previous commits included the modules `plot_pressures.py` and
`orchid/plot-pressures.py`; however, the latter was broken in that it
was copied from `orchid/plot_trajectories.py`.

This commit changes the code to correctly plot the pressure curves for
the loaded project.

This commit is filled with **implicit** hard-coding. For example, we
plot the pressures in a figure with a 2x2 array of subplots. This choice
allows us to begin with working code and to iterate over that code to
repair issues and to add additional features.