Add TODO to rename pressure to monitor pressure
Previous commits used the term "pressure" for pressure measured by a
gauge on a montor well. The previous few commits introduce a different
pressure: the pressure of the fluid injected into a well to treat a
specific stage.

This commit adds a comment suggesting we better distinguish these two
pressures.