Reduce interface "size" by general `unit` method
Previous commits exposed `Project.length_unit` and
`Project.pressure_unit`. The code for these methods was very similar.

This commit introduces `Project.unit(physical_quantity: str) -> str` and
replaced all calls to the unit-specific functions with this more generic
method.

The primary motivation was to make the additional of units for slurry
rate and proppant concentration.