Skeleton measurement and net_measurement
Previous commits put unit conversions either in individual modules
themselves or in the `stub_net` module.

This commit moves some of that work to the `measurement` and
`net_measurement` modules.

Note that this commit *does not* find all the unit conversions in
existing code. The author expects that other future changes will make
that change as part of its work.