Conversion to same unit
Previous commits offered a conversion from Python Measurement instances
to .NET UnitsNet.Unit instances; however, it required the client
developer to supply a target unit even if the unit was the same as the
Measurement unit.

This commit offers to conversion methods: one for the same Measurement
unit and one for a different unit.

Additionally, this commit renames the conversion methods to better
communicate their intent.