Introduce transformed_dom_property
Previous commits only allowe us to create a Python property to access a
DOM property if the code returned the value of the DOM property
unchanged.

This commit introduces `transformed_dom_property` which performs the
same work as `dom_property()` but then executes a transformation
function on that .NET DOM value - most typically but *not* exclusively
to return a Python-specific type.