Fix issues running from installed directory
Poetry requires specifiy both a module *and* a function to invoke when
running a "script." Previous commits specified the `main` function of
the two utility modules; however, argument parsing was handled *before*
`main` was invoked (in the `if __name__ == '__main__'` block).

This commit moves the argument parsing into the module `main` function.
Additionally, this commit supports running from *both* the command line
(in the development directory) and from the poetry "script."