Replace sed-like replacement
Previous commits used a Python "port" of the Unix utility, sed, to
replace the stub path in example notebooks with a user-specified path.

This commit changes that implementation to use string replacement
instead of a regular expression and repairs a number of issues (errors)
with replacing text involving backslashes (Windows-paths, sigh).