(BROKE) Rename method to better reflect its intent
Note that the acceptance / integration tests are broken by this change.

Previous commits used the term "unit" to refer to the Python string
containing the unit abbreviation. Now that I have created the UnitSystem
class (and its concrete children), this usage is no longer a clear
convention.

This commit begins the process of reconciling these two conflicting
usages of the term "unit" by renaming Project.unit to
Project.unit_abbreviation. It is the first of several related changes.