(BROKE) Rename method to better reflect its intent
Note that the acceptance / integration tests are broken by this change.

Previous commits used the term "unit" to refer to the Python string
containing the unit abbreviation. Now that I have created the UnitSystem
class (and its concrete children), this usage is no longer a clear
convention.

This commit is the second step in the initial process of reconciling
these two conflicting usages of the term "unit." It is a close follow-up
of the previous commit by renaming the method `project_units.unit` to
`project_units.unit_abbreviation` to clarify the intent of this method.
Again, it is a step in a sequence of related changes.