Work around issue with mocks
Previous commits specified the interface of mocks using expressions like
`spec=IStage`. This expression ensures that the mocking mechanism **will
not** allow one to mistype an interface method; however, something in
this mechanism raises an error like 'access must occur from class
instance.'

This commit detects the resulting raised exception (`TypeError`) and
then creates the mock **without** the `spec=I...` argument. Removing
this argument does not affect the current code, but increases the
likelihood of introducing an error in the future. (This error occurs
because a change in the .NET API will no longer be detected by unit
tests but only by acceptance tests.)