Workaround UnitsNet Power issue
This commit extends some of the `net_quantity` unit coverage by adding
unit tests for `orchid.physical_quantity.POWER`.

This extension uncovered an issue. UnitsNet defines the `Power` unit
such that the `Value` property returns a value of type `Decimal` instead
of type `double` (like almost all other units). Consequently,
Python.NET, when returning this value to the Python runtime **does not**
return a value of type `float` but actually a .NET `Decimal` value.

This work around converts the .NET `Decimal` value to a Python `float`
value if the physical quantity is POWER.