Move as_unit_system to more accessible location
Previous commits put the function, `as_unit_system`, which converted a
.NET `UnitSystem` to a Python `UnitSystem`, in the `project` module
because it was only needed from that location.

This commit moves that function to `unit_system` to make it accessible
to any code which has access to the .NET `UnitSystem` and to the
`unit_system` module (the latter is true for most modules).
Additionally, since `as_unit_system` returns a `UnitSystem` enumeration
(defined in the `unit_system` module already, this move reduces coupling
between modules.