Move pre-condition to common class and module
Previous commits duplicated the pre-condition code on each of the
methods:
- bottom_location
- center_location
- cluster_location
- top_location

This commit moves that pre-condition to the common method
subsurface_point_in_length_unit. Further, this code moves the code to
test if an instance of `unit_system.UsOilfield` or of
`unit_system.Metric` is actually a LENGTH unit.