Avoid conflicting usages of the term "units."
Previous commits used the name `units` for the single instance of
`pint.UnitRegistry` for the application. This usage conflicted with the
general usage of `units` as an alias for `unit_system`. Although we
support creating `pint` measurements using the application registry, we
do not believe that this use case is the primary use case. Instead, we
will encourage consumers of the Orchid Python API to use the units from
the `unit_system`. We support the option of using `pint` units directly
for either exploratory programming or programmatic conversions to
specific units (known only by the consuming code). If the consuming code
thinks that `registry` is too long for such a common operation, the
consuming code can create its own alias.

This commit renames measurement.units to measurement.registry.