Fix environment error running example scripts
Previous commits run each example script by running the command
`python <example_script.py>`. The result of this is the python
interpreter throwing an exception because it cannot load modules in the
virtual environment.

I do not understand how this occurs.

This commit changes the command to use `sys.executable` to specifically
use the python executable used to run `run_script_examples` (which is
the Python installed in the virtual environment when testing).