"Sentinel" script
Previous commits configured multiple tests by picking a "last test" and
creating a configuration that ends with that test.

This commit creates a "sentinel" script that simply prints 'Finished'.
Now, combined tests like "All tests" and "All tests and docs" can end
with this sentinel script so that I can more easily insert or remove
tests in the combination.