StageDto accepts different types for stage_parts
Previous commits expected the `stage_parts` member to have type
Iterable[StagePartDto]. However, this sequence does not allow the caller
to mock the .NET values returns by the `IStage.Parts` property.

This commit supports an Iterable of StagePartDtos or of IStagePart. This
allows the test to mock the `IStage.Parts` items to return test-specific
values.