Correct error setting start/stop single part
Previous commits used the same logic to handle a single part and to
handle many parts. The integration tests revealed an error in this
approach. The error results when the new start time is *after* the
original stop time.

This commit distinguishes handling a single part and handling multiple
parts.