Rename class to better communicate intent
Previous commits included the `object_id` attribute in the
`DotNetAdapter` class. This prevents deriving `NativeProjectUserData`
from `DotNetAdapter`; thereby, preventing the use of all our work on
adapters.

This commit renames this class in preparation for separating the class
offering the `object_id` attribute and classes which adapt .NET classes
but *do not* have an `object_id` attribute.