Move CreateStageDto and tests to NativeWellAdapter
Previous commits put these classes in native_stage_adapter; however, the
author noticed he need not require consumers of this class to issue two
calls to add a stage (a "create" call and an "add stage" call), but
could simply supply a `CreateStageDto` instance to the `add_stage` call.
This change would return a stage **already** connected to its parent
well instead of using a second step. This approach is the approach used
by the Orchid .NET API.

This commit moves this class and its tests.