Better communicate intent of script
Previous commits used the name `add_stages.py` for the low-level example
of adding one or more stages to a project. When we only offered the
low-level API, this choice was reasonable.

This commit renames this script to better communicate that it uses the
low-level API. Additionally, this renaming frees up the name
`add_stages.py` for an example of adding one or more stages using the
high-level API. Further, this names is more consistent with other names
that use the convention: things using the high-level API are "unadorned"
but things using the low-level API are more clearly labeled as such.