Correct error testing native_well_adapter
Previous commits changed `locations_for_md_kb_values` to pass the
`value` member of the `well_reference_frame` and the `depth_datum`;
however, these commits **did not** change the mock used in testing to
compare the corresponding `value` of the enumeration members. Instead
the mock compared the `value` enumeration members to the members
themselves. (And so, of course, failed to find a match.)

This commit changes the code configuring the mock to compare the
corresponding .NET Enum with the `value` of the enumeration members,
`well_reference_frame`, nd `depth_datum`.